fix js pagination, fix route
