+++ config/routes.yaml
@@ -1,3 +1,3 @@
 controllers:
-    resource: ../src/Controller/
-    type: attribute
+  resource: ../src/Controller/
+  type: attribute
