+++ templates/admin/users/addUserFromAdmin.html.twig
@@ -1,11 +1,9 @@
-{% extends '/admin/adminBase.html.twig' %}
+{% use '/admin/adminBase.html.twig' %}
 
 {% block title %}Créer un nouvel utilisateur
 {% endblock %}
 
 {% block body %}
-	<h1>Créer un nouvel utilisateur</h1>
-
 	{{ form_start(form) }}
 	{{ form_row(form.email) }}
 	{{ form_row(form.roles) }}
@@ -13,8 +11,4 @@
 	{{ form_row(form.firstname) }}
 	{{ form_row(form.lastname) }}
 	{{ form_row(form.company) }}
-
-
-	<button type="submit">Créer</button>
-	{{ form_end(form) }}
 {% endblock %}
