+++ templates/admin/users/updateFromAdmin.html.twig
@@ -3,9 +3,6 @@
 {% block title %}Modifier un utilisateur
 {% endblock %}
 
-{% block javascripts %}
-	{{ encore_entry_script_tags('admin_users') }}
-{% endblock %}
 
 {% block body %}
 	<h1>Modifier un utilisateur</h1>
