+++ templates/register_owner/index.html.twig
@@ -1,0 +1,18 @@
+{% use '/base.html.twig' %}
+
+{% block title %}
+	Hello RegisterOwnerController!
+{% endblock %}
+
+{% block body %}
+	<h1>Créer votre compte:</h1>
+
+	{{ form_start(registrationForm) }}
+	{{ form_row(registrationForm.email) }}
+	{{ form_row(registrationForm.password) }}
+	{{ form_row(registrationForm.firstname) }}
+	{{ form_row(registrationForm.lastname) }}
+
+	<button type="submit">Créer</button>
+	{{ form_end(registrationForm) }}
+{% endblock %}
