--- templates/registration/register.html.twig
@@ -1,21 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Register
-{% endblock %}
-
-{% block body %}
-	<h1>Register</h1>
-
-	{{ form_errors(registrationForm) }}
-
-	{{ form_start(registrationForm) }}
-	{{ form_row(registrationForm.email) }}
-	{{ form_row(registrationForm.plainPassword, {
-            label: 'Password'
-        }) }}
-	{{ form_row(registrationForm.agreeTerms) }}
-	{{ form_row(registrationForm.firstname) }}
-	{{ form_row(registrationForm.lastname) }}
-	<button type="submit" class="btn">Register</button>
-	{{ form_end(registrationForm) }}
-{% endblock %}
